Bug fixes - Engine now no longer freezes after failing to start-up, animations do not lose their frame offsets when the animation ends, unless a new one starts or is manually cleared.

+++ V0.10/drawScreen.c
@@ -1068,7 +1068,7 @@ void DisplayDebugInfo(Camera renderCamera, World *GameWorld, SDL_Renderer *Scree
 		} break;
 	}
 
-	snprintf(text, DEBUG_TEXT_MAX_LENGTH, "%s \nTick: %llu", LEMON_VERSION, TickNumber());
+	snprintf(text, DEBUG_TEXT_MAX_LENGTH, "%s \nTick: %llu", LEMON_VERSION, (long long unsigned int)TickNumber());
 	AddDebugText(text, 16 - (renderCamera.width >> 1), 48 - (renderCamera.height >> 1), 0, DTFORMAT_SCREEN_RELATIVE);
 
 	return;
